[Refactor] Move custom ui components and update imports
