+++ src/pages/Auth/Login.tsx
@@ -1,7 +1,7 @@
 import React from "react";
-import { Header } from "@/components/sections/Header";
+import { Header } from "@/components/ui/custom/core/Header";
 import LoginCard from "@/components/layouts/Auth/LoginCard"; // <- bez { }
-import { Footer } from "@/components/sections/Footer";
+import { Footer } from "@/components/ui/custom/core/Footer";
 
 const Login: React.FC = () => {
     return (
