+++ src/pages/Auth/Signup.tsx
@@ -1,7 +1,7 @@
 import React from "react";
-import { Header } from "@/components/sections/Header";
+import { Header } from "@/components/ui/custom/core/Header";
 import SignupCard from "@/components/layouts/Auth/SignupCard"; // <- bez { }
-import { Footer } from "@/components/sections/Footer";
+import { Footer } from "@/components/ui/custom/core/Footer";
 
 const Signup: React.FC = () => {
     return (
