+++ src/pages/Gamemodes/Gamemodes.tsx
@@ -1,7 +1,7 @@
 import React from "react";
-import { Header } from "@/components/sections/Header";
+import { Header } from "@/components/ui/custom/core/Header";
 import { List } from "@/components/layouts/Gamemodes/List/List";
-import { Footer } from "@/components/sections/Footer";
+import { Footer } from "@/components/ui/custom/core/Footer";
 
 const Gamemodes = () => {
     return (
