+++ src/pages/Landing/Children/EarlyAccess/EarlyAccess.tsx
@@ -1,7 +1,7 @@
 import React from "react";
-import { Header } from "@/components/sections/Header";
+import { Header } from "@/components/ui/custom/core/Header";
 import Preview from "@/components/layouts/Landing/Children/EarlyAccess/Preview";
-import { Footer } from "@/components/sections/Footer";
+import { Footer } from "@/components/ui/custom/core/Footer";
 
 const EarlyAccess = () => {
     return (
