+++ src/pages/Landing/Landing.tsx
@@ -1,10 +1,10 @@
 import React from "react";
-import { Header } from "@/components/sections/Header";
+import { Header } from "@/components/ui/custom/core/Header";
 import { Hero } from "@/components/layouts/Landing/Hero/Hero";
 import { Frame } from "@/components/layouts/Landing/Frame";
 import { TechShowcase } from "@/components/layouts/Landing/TechShowcase";
 import { GameEngine } from "@/components/layouts/Landing/GameEngine";
-import { Footer } from "@/components/sections/Footer";
+import { Footer } from "@/components/ui/custom/core/Footer";
 
 const Landing = () => {
     return (
