+++ src/pages/Servers/Children/Hosting.tsx
@@ -1,6 +1,6 @@
 import React from "react";
-import { Header } from "@/components/sections/Header";
-import { Footer } from "@/components/sections/Footer";
+import { Header } from "@/components/ui/custom/core/Header";
+import { Footer } from "@/components/ui/custom/core/Footer";
 import underconstr from "@/assets/icons/underconstr.png";
 
 const Hosting = () => {
