+++ src/pages/Servers/Servers.tsx
@@ -1,8 +1,8 @@
 import React from "react";
-import { Header } from "@/components/sections/Header";
+import { Header } from "@/components/ui/custom/core/Header";
 import Fetching from "@/components/layouts/Servers/Fetching/Fetching";
 import YourServer from "@/components/layouts/Servers/YourServer";
-import { Footer } from "@/components/sections/Footer";
+import { Footer } from "@/components/ui/custom/core/Footer";
 
 const Servers = () => {
     return (
